address all the comments
