+++ kyuubi-server/src/main/scala/org/apache/kyuubi/server/KyuubiServer.scala
@@ -19,7 +19,11 @@ package org.apache.kyuubi.server
 
 import scala.util.Properties
 
+import org.apache.curator.utils.ZKPaths
 import org.apache.hadoop.security.UserGroupInformation
+import org.apache.zookeeper.CreateMode.PERSISTENT
+import org.apache.zookeeper.KeeperException
+import org.apache.zookeeper.KeeperException.NodeExistsException
 
 import org.apache.kyuubi._
 import org.apache.kyuubi.config.KyuubiConf
@@ -27,6 +31,7 @@ import org.apache.kyuubi.config.KyuubiConf.{FRONTEND_PROTOCOLS, FrontendProtocol
 import org.apache.kyuubi.config.KyuubiConf.FrontendProtocols._
 import org.apache.kyuubi.ha.HighAvailabilityConf._
 import org.apache.kyuubi.ha.client.{ServiceDiscovery, ZooKeeperAuthTypes}
+import org.apache.kyuubi.ha.client.ZooKeeperClientProvider._
 import org.apache.kyuubi.metrics.{MetricsConf, MetricsSystem}
 import org.apache.kyuubi.service.{AbstractBackendService, AbstractFrontendService, Serverable}
 import org.apache.kyuubi.util.{KyuubiHadoopUtils, SignalRegister}
@@ -42,6 +47,35 @@ object KyuubiServer extends Logging {
       zkServer.start()
       conf.set(HA_ZK_QUORUM, zkServer.getConnectString)
       conf.set(HA_ZK_AUTH_TYPE, ZooKeeperAuthTypes.NONE.toString)
+    } else {
+      // create chroot path if necessary
+      val connectionStr = conf.get(HA_ZK_QUORUM)
+      val chrootIndex = connectionStr.indexOf("/")
+      val chrootOption = {
+        if (chrootIndex > 0) Some(connectionStr.substring(chrootIndex))
+        else None
+      }
+      chrootOption.foreach { chroot =>
+        val zkConnectionForChrootCreation = connectionStr.substring(0, chrootIndex)
+        val overrideQuorumConf = conf.clone.set(HA_ZK_QUORUM, zkConnectionForChrootCreation)
+        withZkClient(overrideQuorumConf) { zkClient =>
+          if (zkClient.checkExists().forPath(chroot) == null) {
+            val chrootPath = ZKPaths.makePath(null, chroot)
+            try {
+              zkClient
+                .create()
+                .creatingParentsIfNeeded()
+                .withMode(PERSISTENT)
+                .forPath(chrootPath)
+            } catch {
+              case _: NodeExistsException => // do nothing
+              case e: KeeperException =>
+                throw new KyuubiException(s"Failed to create chroot path '$chrootPath'", e)
+            }
+          }
+        }
+        info(s"Created zookeeper chroot path $chroot")
+      }
     }
 
     val server = new KyuubiServer()
